requirements.txt
langchain>=0.1.16
langchain-google-genai>=0.0.6
openai>=1.0.0
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn[standard]==0.22.0
websockets==11.0.2
Jinja2==3.1.2
